[28] Refact stars tests
